requirements.txt
groq
pandas
streamlit
python-dotenv
fastapi
uvicorn
requests
